Rename the main package to 'skcurve'
We rename the package to avoid the conflicts with another packages and to better naming with considering scikit packages naming.
