+++ skcurve/_distance.py
