+++ skcurve/_numeric.py
