+++ skcurve/_smooth.py
@@ -15,7 +15,7 @@ import scipy.signal.windows as windows
 import scipy.ndimage as ndimage
 
 if ty.TYPE_CHECKING:
-    from curve._base import Curve  # noqa
+    from skcurve._base import Curve  # noqa
 
 
 _smooth_methods = {}  # type: ty.Dict[str, 'SmoothingMethod']
