+++ skcurve/_version.py
